pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-agent"
version = "0.1.0"
description = "Desktop automation agent based on mss, pyautogui, and OpenAI-compatible APIs"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "mss>=9.0.1",
  "pyautogui>=0.9.54",
  "pyperclip>=1.9.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
desktop-agent = "desktop_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
